refactor: 역할에 따른 일관성을 위한 이름, 메서드 명 변경
